requirements.txt
Django==3.0.4
psycopg2-binary==2.8.4
django-extensions==2.2.9
djangorestframework==3.11.0
pytz==2019.3
ipinfo==3.0.0
django-crispy-forms==1.9.0

# ipython==7.12.0
# sqlparse==0.3.0
